+++ Sources/Core/ModernAVPlayer.swift
@@ -122,12 +122,12 @@ public final class ModernAVPlayer: NSObject, ModernAVPlayerExposable {
     }
    
     ///
-    /// Replaces the current item metadata with the new metadata
+    /// Replaces the current item metadata and updates now playing info 
     ///
     /// - parameter metadata: metadata to load
     ///
-    public func updateNowPlayingInfo(metadata: PlayerMediaMetadata) {
-        context.updateNowPlayingInfo(metadata: metadata)
+    public func updateMetadata(_ metadata: PlayerMediaMetadata) {
+        context.updateMetadata(metadata)
     }
     
     /// Begins playback of the current item